Add optional erase pixels based on mask step to GM
This is distinct from cloud erase logic as, this happens per dataset before any
fusing.

This is to deal with masks, like fmask, that might not be defined for every
valid pixel in the input. So you might have pixels that have valid data in the
data bands, but for which no valid cloud classification exists.