Move odc.geom out of the way
This library was used to prototype data loading from COG overviews, this
functionality has been moved to datacube. There are still some leftover code in
there that might be useful in the future, but for now it's not being used.

Moving it out of libs/ means it won't be installed for testing, since there are
no tests in there anyway it's a win.