Fix metrics cardinality explosion
Metrics are not logs. Pods are short lived. If we ask statsd to keep tracking a separate set of metrics for every unique pod name it has ever encountered, then we will accumulate tens of thousands of separate metrics over the course of a week, all of which being actively exported to prometheus. This impacts the performance and cost of monitoring.