Fix GHA: Use postgresql service
Previously, a postgres data directory was being manually created
using the low level postgres admin tools like initdb and pg_ctl.
This broke, because the Ubuntu postgresql package moved those
tools to /usr/lib/postgresql/14/bin.

When I added that to the path, the server still failed to start,
so lets use the more common pattern within GHA, and run Postgresql
as a service, updating the environment variables to use it.